+++ apps/engine/contracts/game/src/lib.rs
@@ -10,6 +10,10 @@ use soroban_ecs::{Component, ComponentTrait};
 use soroban_ecs::{System, SystemParam};
 use soroban_ecs::prelude::*;
 
+mod storage;
+
+use storage::GameWorldData;
+
 #[contracttype]
 #[derive(Clone)]
 pub struct GamePosition(pub u32, pub u32);
@@ -99,37 +103,38 @@ impl CombatSystem {
 // GameWorldContract
 #[contract]
 pub struct GameWorldContract;
-/// Contract data structure to store the game world state
-#[derive(Clone, Debug)]
-pub struct GameWorldData {
-    /// Flag indicating if the contract has been initialized
-    pub is_initialized: bool,
-    /// Number of entities in the world
-    pub entity_count: u32,
-    // Number of dead entities in the world
-    pub dead_entity: u32,
-    /// World data
-    pub world_data: World,
-}
-// Manual implementation of Soroban traits for GameWorldData
-impl IntoVal<Env, soroban_sdk::Val> for GameWorldData {
-    fn into_val(&self, env: &Env) -> soroban_sdk::Val {
-        let data = (self.is_initialized, self.entity_count, self.dead_entity);
-        data.into_val(env)
-    }
-}
-impl TryFromVal<Env, soroban_sdk::Val> for GameWorldData {
-    type Error = soroban_sdk::ConversionError;
-    fn try_from_val(env: &Env, val: &soroban_sdk::Val) -> Result<Self, Self::Error> {
-        let (is_initialized, entity_count, dead_entity): (bool, u32, u32) = TryFromVal::try_from_val(env, val)?;
-        Ok(GameWorldData {
-            is_initialized,
-            entity_count,
-            dead_entity,
-            world_data: World::new(),
-        })
-    }
-}
+
+// /// Contract data structure to store the game world state
+// #[derive(Clone, Debug)]
+// pub struct GameWorldData {
+//     /// Flag indicating if the contract has been initialized
+//     pub is_initialized: bool,
+//     /// Number of entities in the world
+//     pub entity_count: u32,
+//     // Number of dead entities in the world
+//     pub dead_entity: u32,
+//     /// World data
+//     pub world_data: World,
+// }
+// // Manual implementation of Soroban traits for GameWorldData
+// impl IntoVal<Env, soroban_sdk::Val> for GameWorldData {
+//     fn into_val(&self, env: &Env) -> soroban_sdk::Val {
+//         let data = (self.is_initialized, self.entity_count, self.dead_entity);
+//         data.into_val(env)
+//     }
+// }
+// impl TryFromVal<Env, soroban_sdk::Val> for GameWorldData {
+//     type Error = soroban_sdk::ConversionError;
+//     fn try_from_val(env: &Env, val: &soroban_sdk::Val) -> Result<Self, Self::Error> {
+//         let (is_initialized, entity_count, dead_entity): (bool, u32, u32) = TryFromVal::try_from_val(env, val)?;
+//         Ok(GameWorldData {
+//             is_initialized,
+//             entity_count,
+//             dead_entity,
+//             world_data: World::new(),
+//         })
+//     }
+// }
 
 /// Initialize the contract with an empty ECS world
 #[contractimpl]
@@ -144,12 +149,14 @@ impl GameWorldContract {
             dead_entity:0,
             world_data: World::new(),
         };
-        Self::save_contract_data(env, &data);
+        // Self::save_contract_data(env, &data);
+        storage::save_contract_data(env, &data);
         data
     }
     /// Create a new entity with a position component
     pub fn spawn_entity(env: &Env, x: u32, y: u32) -> u32 {
-        let mut contract_data = Self::get_contract_data(env);
+        // let mut contract_data = Self::get_contract_data(env);
+        let mut contract_data = storage::get_contract_data(env);
         // Create a new ECS world
         let _world = World::new();
         // Create a position component using the ECS system
@@ -160,17 +167,21 @@ impl GameWorldContract {
         let health = Health(100);
         let entity_data: (u32, u32, u32, u32) = (entity_id, position.0, position.1, health.0.try_into().unwrap());
         let val: soroban_sdk::Val = entity_data.into_val(env);
-        env.storage().instance().set(&entity_key, &val);
+        // env.storage().instance().set(&entity_key, &val);
+        storage::set_entity_data(env, val);
         // Update entity count
         contract_data.entity_count += 1;
-        Self::save_contract_data(env, &contract_data);
+        // Self::save_contract_data(env, &contract_data);
+        storage::save_contract_data(env, &contract_data);
         entity_id
     }
 
     // Move an entity by the given delta using the MovementSystem
     pub fn move_entity(env: &Env, entity_id: u32, dx: i32, dy: i32) -> bool {
         let entity_key = symbol_short!("entity");
-        if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        // if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        if let Some(entity_data) = storage::get_entity_data(&env) {
+        
             if let Ok((id, x, y, health)) = <(u32, u32, u32, u32)>::try_from_val(env, &entity_data) {
                 if id == entity_id {
                     let current_position = GamePosition ( x, y );
@@ -178,7 +189,9 @@ impl GameWorldContract {
                     // Store updated position
                     let updated_entity_data: (u32, u32, u32, u32) = (id, new_position.0, new_position.1, health);
                     let val: soroban_sdk::Val = updated_entity_data.into_val(env);
-                    env.storage().instance().set(&entity_key, &val);
+                    // env.storage().instance().set(&entity_key, &val);
+                    storage::set_entity_data(env, val);
+
                     return true;
                 }
             }
@@ -188,9 +201,9 @@ impl GameWorldContract {
 
     /// Attack entity
     pub fn attack_entity(env: &Env, entity_id: u32) -> bool {
-        
-        let entity_key = symbol_short!("entity");
-        if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        // let entity_key = symbol_short!("entity");
+        // if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        if let Some(entity_data) = storage::get_entity_data(&env) {
             if let Ok((id, x, y, health)) = <(u32, u32, u32, u32)>::try_from_val(env, &entity_data) {
                 if id == entity_id {
                     let current_heath = Health(health.try_into().unwrap());
@@ -198,14 +211,17 @@ impl GameWorldContract {
                     if new_health > 0 {
                         let updated_entity_data: (u32, u32, u32, u32) = (id, x, y, new_health.try_into().unwrap());
                         let val: soroban_sdk::Val = updated_entity_data.into_val(env);
-                        env.storage().instance().set(&entity_key, &val);
+                        // env.storage().instance().set(&entity_key, &val);
+                        storage::set_entity_data(env, val);
                         return true;
                     } else {
-                        let mut contract_data = Self::get_contract_data(env);
-                        Self::despawn_entity(env, entity_id);
+                        // let mut contract_data = Self::get_contract_data(env);
+                        let mut contract_data = storage::get_contract_data(env);
+                        // Self::despawn_entity(env, entity_id);
                         contract_data.dead_entity += 1;
-                        Self::save_contract_data(env, &contract_data);
-                        true;
+                        // Self::save_contract_data(env, &contract_data);
+                        storage::save_contract_data(env, &contract_data);
+                        return true;
                     }
                 }
             }
@@ -213,10 +229,11 @@ impl GameWorldContract {
         false
     }
 
-    /// Get entity position
+    // Get entity position
     pub fn get_entity_position(env: &Env, entity_id: u32) -> Option<GamePosition> {
-        let entity_key = symbol_short!("entity");
-        if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        // let entity_key = symbol_short!("entity");
+        // if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        if let Some(entity_data) = storage::get_entity_data(&env) {
             if let Ok((id, x, y, _)) = <(u32, u32, u32, u32)>::try_from_val(env, &entity_data) {
                 if id == entity_id {
                     return Some(GamePosition ( x, y ));
@@ -227,8 +244,9 @@ impl GameWorldContract {
     }
 
     pub fn get_entity_health(env: &Env, entity_id: u32) -> Option<Health> {
-        let entity_key = symbol_short!("entity");
-        if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        // let entity_key = symbol_short!("entity");
+        // if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        if let Some(entity_data) = storage::get_entity_data(&env) {
             if let Ok((id, x, y, health)) = <(u32, u32, u32, u32)>::try_from_val(env, &entity_data) {
                 if id == entity_id {
                     return Some(Health ( health.try_into().unwrap() ));
@@ -239,30 +257,36 @@ impl GameWorldContract {
     }
     /// Get the total number of entities in the world
     pub fn entity_count(env: &Env) -> u32 {
-        let contract_data = Self::get_contract_data(env);
+        // let contract_data = Self::get_contract_data(env);
+        let contract_data = storage::get_contract_data(env);
         contract_data.entity_count
     }
 
      /// Get the total number of dead entities in the world
     pub fn dead_entity_count(env: &Env) -> u32 {
-        let contract_data = Self::get_contract_data(env);
+        // let contract_data = Self::get_contract_data(env);
+        let contract_data = storage::get_contract_data(env);
         contract_data.dead_entity
     }
 
     // /// Remove an entity from the world
     pub fn despawn_entity(env: &Env, entity_id: u32) -> bool {
-        let entity_key = symbol_short!("entity");
+        // let entity_key = symbol_short!("entity");
         // Check if entity exists
-        if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        // if let Some(entity_data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&entity_key) {
+        if let Some(entity_data) = storage::get_entity_data(&env) {
             if let Ok((id, _, _, _)) = <(u32, u32, u32, u32)>::try_from_val(env, &entity_data) {
                 if id == entity_id {
                     // Remove entity by clearing storage
-                    env.storage().instance().remove(&entity_key);
+                    // env.storage().instance().remove(&entity_key);
+                    storage::remove_entity_data(&env);
                     // Update entity count
-                    let mut contract_data = Self::get_contract_data(env);
+                    // let mut contract_data = Self::get_contract_data(env);
+                    let mut contract_data = storage::get_contract_data(env);
                     if contract_data.entity_count > 0 {
                         contract_data.entity_count -= 1;
-                        Self::save_contract_data(env, &contract_data);
+                        // Self::save_contract_data(env, &contract_data);
+                        storage::save_contract_data(env, &contract_data);
                     }
                     return true;
                 }
@@ -272,31 +296,31 @@ impl GameWorldContract {
     }
 }
 
-impl GameWorldContract {
-    /// Get contract data from storage
-    fn get_contract_data(env: &Env) -> GameWorldData {
-        let key = symbol_short!("contract");
-        if let Some(data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&key) {
-            GameWorldData::try_from_val(env, &data).unwrap_or_else(|_| GameWorldData {
-                is_initialized: false,
-                entity_count: 0,
-                dead_entity: 0,
-                world_data: World::new(),
-            })
-        } else {
-            // Initialize with empty world if no data exists
-            GameWorldData {
-                is_initialized: false,
-                entity_count: 0,
-                dead_entity: 0,
-                world_data: World::new(),
-            }
-        }
-    }
-    /// Save contract data to storage
-    fn save_contract_data(env: &Env, data: &GameWorldData) {
-        let key = symbol_short!("contract");
-        let val: soroban_sdk::Val = data.into_val(env);
-        env.storage().instance().set(&key, &val);
-    }
-}
+// impl GameWorldContract {
+//     /// Get contract data from storage
+//     fn get_contract_data(env: &Env) -> GameWorldData {
+//         let key = symbol_short!("contract");
+//         if let Some(data) = env.storage().instance().get::<soroban_sdk::Symbol, soroban_sdk::Val>(&key) {
+//             GameWorldData::try_from_val(env, &data).unwrap_or_else(|_| GameWorldData {
+//                 is_initialized: false,
+//                 entity_count: 0,
+//                 dead_entity: 0,
+//                 world_data: World::new(),
+//             })
+//         } else {
+//             // Initialize with empty world if no data exists
+//             GameWorldData {
+//                 is_initialized: false,
+//                 entity_count: 0,
+//                 dead_entity: 0,
+//                 world_data: World::new(),
+//             }
+//         }
+//     }
+//     /// Save contract data to storage
+//     fn save_contract_data(env: &Env, data: &GameWorldData) {
+//         let key = symbol_short!("contract");
+//         let val: soroban_sdk::Val = data.into_val(env);
+//         env.storage().instance().set(&key, &val);
+//     }
+// }
